pkg/logging: move pkg/debugdetection to pkg/logging
Since the debug detection is only used for the log package it doesn't
make sense to have them split in different packages.
