+++ pkg/logging/logfields/helpers.go
@@ -16,9 +16,6 @@ package logfields
 
 import (
 	"fmt"
-
-	// Initialize logrus package with debug log level if needed
-	_ "github.com/cilium/cilium/pkg/debugdetection"
 )
 
 // Repr formats an object with the Printf %+v formatter
